ARTEMIS-4558 Idempotent Mirrored ACKs
Mirror acks should be performed atomically with the storage of the source ACK. Both the send of the ack and the recording of the ack should be part of the same transaction (in case of transactional).

We are also adding support on transactions for an afterWired callback for the proper plug of OperationContext sync.

(cherry picked from commit e7a1a6df57c67071454a33de0938ff0cf6329a61)

downstream: ENTMQBR-8633